LPD-44159 Reorganize tests according to the new FDSSerializer class hierarchy
